backend/requirements.txt
Flask==3.0.3
Flask-CORS==3.0.10
Flask-Mail==0.9.1 
Flask-Mailman==1.1.1
requests==2.28.1
python-dotenv==0.20.0
Werkzeug==3.0.3
Jinja2==3.1.4
MarkupSafe==3.0.1
flask-swagger-ui==4.11.1
